--- styles/i.html
@@ -1,47 +1,0 @@
-<li class="list__item">
-  <div class="list__checkbox">checkbox</div>
-  <div class="list__item-body">
-    <div class="list__item-content">content</div>
-    <ul class="list__item-child">
-      <li class="list__item">
-        <div class="list__checkbox">checkbox2</div>
-        <div class="list__item-body">
-          <div class="list__item-content">content3</div>
-          <ul class="list__item-child"></ul>
-        </div>
-      </li>
-      <li class="list__item">
-        <div class="list__checkbox">checkbox3</div>
-        <div class="list__item-body">
-          <div class="list__item-content">content3</div>
-          <ul class="list__item-child"></ul>
-        </div>
-      </li>
-    </ul>
-  </div>
-</li>
-
-<style>
-.list__item {
-  display: grid;
-  grid-template-columns: auto 1fr;
-  grid-template-rows: auto auto;
-  grid-template-areas:
-    "checkbox content"
-    ". child";
-}
-
-.list__checkbox {
-  grid-area: checkbox;
-  vertical-align: middle;
-}
-
-.list__item-content {
-  grid-area: content;
-}
-
-.list__item-child {
-  grid-area: child;
-}
-
-</style>
