+++ styles/index.pcss
@@ -8,13 +8,13 @@
   --checkbox-background: #fff;
   --color-border: #C9C9C9;
   --color-bg-checked: #369FFF;
-  --line-height: 1.57em;
+  --line-height: 1.6em;
   --color-bg-checked-hover: #0059AB;
   --color-tick: #fff;
-  --size-checkbox: 20px;
+  --size-checkbox: 1.2em;
 
   &__item {
-    line-height: 1.6em;
+    line-height: var(--line-height);
     display: grid;
     grid-template-columns: auto 1fr;
     grid-template-rows: auto auto;
@@ -68,7 +68,6 @@
   }
 
   &__checkbox {
-    padding-top: 0.1em;
     grid-area: checkbox;
     width: var(--size-checkbox);
     height: var(--size-checkbox);
@@ -95,7 +94,7 @@
     }
 
     &--checked {
-      line-height: 1.6em;
+      line-height: var(--line-height);
 
       @media (hover: hover) {
         &:not(&--no-hover):hover {
